+++ OCCtask1/src/main/java/com/example/OCCtask1/ProjectRepository.java
@@ -2,7 +2,7 @@ package com.example.OCCtask1;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ProjectRepository extends MongoRepository<Project, Long> {
+public interface ProjectRepository extends MongoRepository<Project, String> {
 
 
 }
