+++ OCCtask1/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+
+spring.data.mongodb.host=localhost
+spring.data.mongodb.port=27017
+spring.data.mongodb.database=ProjectsBase
+
+server.port=8080
+spring.datasource.url=jdbc:mongodb://localhost:27017/itproger
+spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
